api-python/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai-whisper==20231117
